+++ requests_cache/backends/dynamodb.py
@@ -25,6 +25,7 @@ class DynamoDbCache(BaseCache):
 
     Args:
         table_name: DynamoDB table name
+        create_table: Whether or not to automatically create the dynamo backend table.
         connection: :boto3:`DynamoDB Resource <services/dynamodb/service-resource/index.html#service-resource>`
             object to use instead of creating a new one
         ttl: Use DynamoDB TTL to automatically remove expired items
@@ -34,6 +35,7 @@ class DynamoDbCache(BaseCache):
     def __init__(
         self,
         table_name: str = 'http_cache',
+        create_table: bool = True,
         *,
         ttl: bool = True,
         connection: Optional[ServiceResource] = None,
@@ -45,6 +47,7 @@ class DynamoDbCache(BaseCache):
         skwargs = {'serializer': serializer, **kwargs} if serializer else kwargs
         self.responses = DynamoDbDict(
             table_name,
+            create_table=create_table,
             ttl=ttl,
             connection=connection,
             decode_content=decode_content,
@@ -59,6 +62,7 @@ class DynamoDbDict(BaseStorage):
 
     Args:
         table_name: DynamoDB table name
+        create_table: Whether or not to automatically create the dynamo backend table.
         connection: :boto3:`DynamoDB Resource <services/dynamodb/service-resource/index.html#service-resource>`
             object to use instead of creating a new one
         ttl: Use DynamoDB TTL to automatically remove expired items
@@ -68,6 +72,7 @@ class DynamoDbDict(BaseStorage):
     def __init__(
         self,
         table_name: str,
+        create_table: bool = True,
         ttl: bool = True,
         connection: Optional[ServiceResource] = None,
         serializer: Optional[SerializerType] = dynamodb_document_serializer,
@@ -82,7 +87,8 @@ class DynamoDbDict(BaseStorage):
         self.ttl = ttl
 
         self._table = self.connection.Table(self.table_name)
-        self._create_table()
+        if create_table:
+            self._create_table()
         if ttl:
             self._enable_ttl()